requirements.txt
python-telegram-bot==21.10
python-dotenv==1.0.1
httpx==0.28.1
pytesseract==0.3.13
PyPDF2==3.0.1
Pillow==11.1.0
SpeechRecognition==3.11.0
langdetect==1.0.9
APScheduler==3.10.4
python-dateutil==2.9.0
